Move api to auth specific and user specific controllers
Move routes that manage the logged-in user's data to the currentUserController file:

  Move GET /users/preferences route to GET /current_user/preferences
  Move POST /users/preferences route to POST /current_user/preferences
Split up userController.js into authController.js (/auth prefix) and usersController.js (/users prefix):

  Move all login/authentication-related routes from /users to /auth
 /users/login to /auth/login
 /users/signup to /auth/signup
 /users/forgot to /auth/forgot
 /users/reset to /auth/reset_password
 /users/confirmation to /auth/confirm
 /users/resendconfirmation to /auth/resend_confirmation
 /users/login/google to /auth/login/google
Combine GET /sketches and GET /sketches/:userName into a single route:

  Delete the GET /sketches route
  Move GET /sketches/:userName route to GET /users/:userName/sketches
